refactor: relocate entitiesview and entityview to shared with per-site side columns (#1609)
