+++ packages/shared/views/EntityView/assembly/components/Side/side.styles.ts
