+++ site-config/brc-analytics/local/index/workflowEntityConfig.ts
@@ -1,7 +1,7 @@
 import { getWorkflowId } from "@/apis/catalog/brc-analytics-catalog/common/utils";
 import { type EntityConfig } from "@databiosphere/findable-ui/lib/config/entities";
 import { EXPLORE_MODE } from "@databiosphere/findable-ui/lib/hooks/useExploreMode/types";
-import { type AppEntityConfig } from "@site-config/common/entities";
+import { type AppEntityConfig } from "@repo/shared/config/types";
 import { CATEGORY_GROUP_CONFIG } from "./workflow/categoryGroupConfig";
 import { COLUMNS } from "./workflow/columns";
 import { TABLE_OPTIONS } from "./workflow/tableOptions";
