+++ site-config/ga2/local/index/workflow/workflowEntityConfig.ts
@@ -1,7 +1,7 @@
 import { getWorkflowId } from "@/apis/catalog/ga2/utils";
 import { type EntityConfig } from "@databiosphere/findable-ui/lib/config/entities";
 import { EXPLORE_MODE } from "@databiosphere/findable-ui/lib/hooks/useExploreMode/types";
-import { type AppEntityConfig } from "@site-config/common/entities";
+import { type AppEntityConfig } from "@repo/shared/config/types";
 import { CATEGORY_GROUP_CONFIG } from "./categoryGroupConfig";
 import { COLUMNS } from "./columns";
 import { TABLE_OPTIONS } from "./tableOptions";
