+++ sites/brc-analytics/views/EntityView/assembly/components/Side/brc/side.styles.ts
