+++ sites/brc-analytics/views/EntityView/assembly/components/Side/brc/side.tsx
@@ -1,20 +1,20 @@
-import { AssemblyDetails } from "@/views/EntityView/assembly/components/Side/ga2/components/AssemblyDetails/AssemblyDetails";
-import { StyledFluidPaper } from "@/views/EntityView/assembly/components/Side/side.styles";
+import {
+  buildAssemblyDetails,
+  buildAssemblyResources,
+  buildOrganismDetails,
+} from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
 import { mapAssemblyToOrganism } from "@/views/WorkflowInputsView/utils";
 import { BackPageContentSideColumn } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/backPageView.styles";
 import { AssemblyFavoriteButton } from "@repo/shared/components/Favorites/components/AssemblyFavoriteButton/assemblyFavoriteButton";
-import {
-  buildAssemblyResources,
-  buildOrganismDetails,
-} from "@repo/shared/viewModelBuilders/viewModelBuilders";
 import { AnalysisPortals } from "@repo/shared/views/EntityView/assembly/components/Side/AnalysisPortals/analysisPortals";
+import { StyledFluidPaper } from "@repo/shared/views/EntityView/assembly/components/Side/side.styles";
 import { KeyValueSection } from "@repo/shared/views/EntityView/components/KeyValueSection/keyValueSection";
 import { type JSX } from "react";
 import { StyledSection } from "./side.styles";
 import { type Props } from "./types";
 
 /**
- * Side column component for the GA2 AnalyzeView, displaying assembly details and resources.
+ * Side column component for the BRC AnalyzeView, displaying assembly details and resources.
  * @param props - Component props.
  * @param props.assembly - Assembly.
  * @returns JSX element representing the side column content.
@@ -30,7 +30,10 @@ export const Side = ({ assembly }: Props): JSX.Element => {
           {...buildOrganismDetails(mapAssemblyToOrganism(assembly))}
           title="Organism Details"
         />
-        <AssemblyDetails assembly={assembly} />
+        <KeyValueSection
+          {...buildAssemblyDetails(assembly)}
+          title="Assembly Details"
+        />
         <AnalysisPortals
           {...buildAssemblyResources(assembly)}
           title="Resources"
