+++ sites/brc-analytics/views/EntityView/assembly/components/Side/brc/types.ts
