+++ sites/brc-analytics/views/PriorityPathogensView/types.ts
@@ -1,7 +1,5 @@
-import {
-  type EntitiesResponse,
-  type Outbreak,
-} from "@/apis/catalog/brc-analytics-catalog/common/entities";
+import { type Outbreak } from "@/apis/catalog/brc-analytics-catalog/common/entities";
+import { type EntitiesResponse } from "@repo/shared/services/staticGeneration/entities/types";
 
 export interface Props {
   data: EntitiesResponse<Outbreak>;
