+++ sites/ga2/views/EntityView/assembly/components/Side/ga2/types.ts
